2020.4.10

+++ qingcheng_service_goods/src/main/java/com/qingcheng/service/impl/SpuServiceImpl.java
@@ -1,4 +1,5 @@
 package com.qingcheng.service.impl;
+
 import com.alibaba.dubbo.config.annotation.Service;
 import com.alibaba.fastjson.JSON;
 import com.github.pagehelper.Page;
@@ -37,10 +38,10 @@ public class SpuServiceImpl implements SpuService {
 
     @Autowired
     private AuditStatusMapper auditStatusMapper;
-    private Spu spu;
 
     /**
      * 返回全部记录
+     *
      * @return
      */
     public List<Spu> findAll() {
@@ -49,18 +50,20 @@ public class SpuServiceImpl implements SpuService {
 
     /**
      * 分页查询
+     *
      * @param page 页码
      * @param size 每页记录数
      * @return 分页结果
      */
     public PageResult<Spu> findPage(int page, int size) {
-        PageHelper.startPage(page,size);
+        PageHelper.startPage(page, size);
         Page<Spu> spus = (Page<Spu>) spuMapper.selectAll();
-        return new PageResult<Spu>(spus.getTotal(),spus.getResult());
+        return new PageResult<Spu>(spus.getTotal(), spus.getResult());
     }
 
     /**
      * 条件查询
+     *
      * @param searchMap 查询条件
      * @return
      */
@@ -71,20 +74,22 @@ public class SpuServiceImpl implements SpuService {
 
     /**
      * 分页+条件查询
+     *
      * @param searchMap
      * @param page
      * @param size
      * @return
      */
     public PageResult<Spu> findPage(Map<String, Object> searchMap, int page, int size) {
-        PageHelper.startPage(page,size);
+        PageHelper.startPage(page, size);
         Example example = createExample(searchMap);
         Page<Spu> spus = (Page<Spu>) spuMapper.selectByExample(example);
-        return new PageResult<Spu>(spus.getTotal(),spus.getResult());
+        return new PageResult<Spu>(spus.getTotal(), spus.getResult());
     }
 
     /**
      * 根据Id查询
+     *
      * @param id
      * @return
      */
@@ -94,6 +99,7 @@ public class SpuServiceImpl implements SpuService {
 
     /**
      * 新增
+     *
      * @param spu
      */
     public void add(Spu spu) {
@@ -102,6 +108,7 @@ public class SpuServiceImpl implements SpuService {
 
     /**
      * 修改
+     *
      * @param spu
      */
     public void update(Spu spu) {
@@ -109,15 +116,49 @@ public class SpuServiceImpl implements SpuService {
     }
 
     /**
-     *  删除
+     * 真删除商品
+     * 如果商品需要物理删除，必须是先逻辑删除才能进行物理删除，删除前需要检查状态。
+     *
      * @param id
      */
-    public void delete(String id) {
+    public void realdelete(String id) {
+        Spu spu = spuMapper.selectByPrimaryKey(id);
+        if (spu == null) {
+            throw new RuntimeException("商品不存在");
+        }
+        if (!"1".equals(spu.getIsDelete())) {
+            throw new RuntimeException("必须先逻辑删除才能进行物理删除");
+        }
         spuMapper.deleteByPrimaryKey(id);
+    }
+
+
+    /**
+     * 假删除商品
+     *
+     * @param id
+     */
+    public void falsedelete(String id) {
+        //逻辑删除商品,修改spu表is_delete字段为1
+        Spu spu = new Spu();
+        spu.setId(id);
+        spu.setIsDelete("1");
+        spuMapper.updateByPrimaryKeySelective(spu);
+    }
+
+    /**
+     * 回收商品，修改spu表is_delete字段为0
+     */
+    public void restore(String id) {
+        Spu spu = new Spu();
+        spu.setId(id);
+        spu.setIsDelete("0");
+        spuMapper.updateByPrimaryKeySelective(spu);
     }
 
     /**
      * 保存商品(新增和修改)
+     *
      * @param goods 商品组合实体类
      */
     @Transactional(rollbackFor = Exception.class)
@@ -128,16 +169,16 @@ public class SpuServiceImpl implements SpuService {
         // ③将值存入数据库
         Spu spu = goods.getSpu();
 
-        if (spu.getId()==null){
+        if (spu.getId() == null) {
             //增加
-            spu.setId(idWorker.nextId()+"");
+            spu.setId(idWorker.nextId() + "");
             spuMapper.insert(spu);
-        }else{
+        } else {
             //修改
             //①删除原来的(spu下的的sku列表)sku列表（这样修改sku表就相当于再一次增加，主要能使其合并）
-            Example example=new Example(Sku.class);
+            Example example = new Example(Sku.class);
             Example.Criteria criteria = example.createCriteria();
-            criteria.andEqualTo("spuId",spu.getId());
+            criteria.andEqualTo("spuId", spu.getId());
             skuMapper.deleteByExample(example);
             //②对spu执行updateByPrimaryKeySelective操作
             //updateByPrimaryKeySelective:如果修改数据有null，则不会被覆盖，而且展示原来的值
@@ -148,10 +189,10 @@ public class SpuServiceImpl implements SpuService {
         Date date = new Date();
         Category category = categoryMapper.selectByPrimaryKey(spu.getCategory3Id());
         //保存sku列表的信息
-        for (Sku sku:skus){
-            if (sku.getId()==null){
+        for (Sku sku : skus) {
+            if (sku.getId() == null) {
                 //增加
-                sku.setId(idWorker.nextId()+"");
+                sku.setId(idWorker.nextId() + "");
                 //创建日期
                 sku.setCreateTime(date);
             }
@@ -162,16 +203,16 @@ public class SpuServiceImpl implements SpuService {
 
             String skuSpec = sku.getSpec();
             //没有规格（spec）属性，我们要对其进行判断，避免因空值产生
-            if(skuSpec==null||"".equals(skuSpec)){
+            if (skuSpec == null || "".equals(skuSpec)) {
                 sku.setSpec("{}");
             }
 
-            //将json转换成map
+            //将json转换成对象map
             //json : {"颜色":"绿","机身内存":"8G"}
             //map:{key:颜色；value：绿}
-            Map<String ,String> skuSpecMap = JSON.parseObject(skuSpec, Map.class);
-            for (String value:skuSpecMap.values()){
-                name+=" "+value;
+            Map<String, String> skuSpecMap = JSON.parseObject(skuSpec, Map.class);
+            for (String value : skuSpecMap.values()) {
+                name += " " + value;
             }
             //sku名称
             sku.setName(name);
@@ -194,7 +235,7 @@ public class SpuServiceImpl implements SpuService {
             categoryBrand.setCategoryId(spu.getCategory3Id());
             //注意问题：反复添加主键重复，因此加一个判断条件(判断是否有这条数据)
             int count = categoryBrandMapper.selectCount(categoryBrand);
-            if (count==0){
+            if (count == 0) {
                 categoryBrandMapper.insert(categoryBrand);
             }
         }
@@ -203,6 +244,7 @@ public class SpuServiceImpl implements SpuService {
 
     /**
      * 查询商品id
+     *
      * @param id
      */
     @Transactional(rollbackFor = Exception.class)
@@ -215,9 +257,9 @@ public class SpuServiceImpl implements SpuService {
           from tb_sku
           where spu_id='1247483467168616448'
         */
-        Example example=new Example(Sku.class);
+        Example example = new Example(Sku.class);
         Example.Criteria criteria = example.createCriteria();
-        criteria.andEqualTo("spuId",id);
+        criteria.andEqualTo("spuId", id);
         List<Sku> skuList = skuMapper.selectByExample(example);
         //封装为Goods
         Goods goods = new Goods();
@@ -228,6 +270,7 @@ public class SpuServiceImpl implements SpuService {
 
     /**
      * 商品审核(audit)
+     *
      * @param id,status,message
      */
     public void audit(String id, String status, String message) {
@@ -237,7 +280,7 @@ public class SpuServiceImpl implements SpuService {
         Spu spu = new Spu();
         spu.setId(id);
         spu.setStatus(status);
-        if ("1".equals(status)){
+        if ("1".equals(status)) {
             spu.setIsMarketable("1");
         }
         spuMapper.updateByPrimaryKeySelective(spu);
@@ -273,7 +316,7 @@ public class SpuServiceImpl implements SpuService {
         //1、商品上架
         //需要验证该商品是否审核通过，未审核通过的商品不能上架。
         Spu spu = spuMapper.selectByPrimaryKey(id);
-        if (!"1".equals(spu.getStatus())){
+        if (!"1".equals(spu.getStatus())) {
             throw new RuntimeException("此商品未审核通过，不能上架");
         }
         spu.setIsMarketable("1");
@@ -293,119 +336,120 @@ public class SpuServiceImpl implements SpuService {
         //意思是将满足这些条件的行的is_marketable这一列值全部改为1
         spu.setIsMarketable("1");
         //批量修改
-        Example example=new Example(Spu.class);
+        Example example = new Example(Spu.class);
         Example.Criteria criteria = example.createCriteria();
         //数组转换为集合List ：Arrays.asList(ids)
-        criteria.andIn("id",Arrays.asList(ids));
+        criteria.andIn("id", Arrays.asList(ids));
         //未审核通过的商品不能上架
-        criteria.andEqualTo("status",1);
+        criteria.andEqualTo("status", 1);
         //非删除的
-        criteria.andEqualTo("isDelete",0);
+        criteria.andEqualTo("isDelete", 0);
         //只考虑把下架(0)的商品做上架(1)操作
-        criteria.andEqualTo("isMarketable",0);
+        criteria.andEqualTo("isMarketable", 0);
         //用这个返回值是int
-        int i = spuMapper.updateByExampleSelective(spu,example);
+        int i = spuMapper.updateByExampleSelective(spu, example);
         //2.记录商品日志
         return i;
     }
 
     /**
      * 构建查询条件
+     *
      * @param searchMap
      * @return
      */
-    private Example createExample(Map<String, Object> searchMap){
-        Example example=new Example(Spu.class);
+    private Example createExample(Map<String, Object> searchMap) {
+        Example example = new Example(Spu.class);
         Example.Criteria criteria = example.createCriteria();
-        if(searchMap!=null){
+        if (searchMap != null) {
             // 主键
-            if(searchMap.get("id")!=null && !"".equals(searchMap.get("id"))){
-                criteria.andLike("id","%"+searchMap.get("id")+"%");
+            if (searchMap.get("id") != null && !"".equals(searchMap.get("id"))) {
+                criteria.andLike("id", "%" + searchMap.get("id") + "%");
             }
             // 货号
-            if(searchMap.get("sn")!=null && !"".equals(searchMap.get("sn"))){
-                criteria.andLike("sn","%"+searchMap.get("sn")+"%");
+            if (searchMap.get("sn") != null && !"".equals(searchMap.get("sn"))) {
+                criteria.andLike("sn", "%" + searchMap.get("sn") + "%");
             }
             // SPU名
-            if(searchMap.get("name")!=null && !"".equals(searchMap.get("name"))){
-                criteria.andLike("name","%"+searchMap.get("name")+"%");
+            if (searchMap.get("name") != null && !"".equals(searchMap.get("name"))) {
+                criteria.andLike("name", "%" + searchMap.get("name") + "%");
             }
             // 副标题
-            if(searchMap.get("caption")!=null && !"".equals(searchMap.get("caption"))){
-                criteria.andLike("caption","%"+searchMap.get("caption")+"%");
+            if (searchMap.get("caption") != null && !"".equals(searchMap.get("caption"))) {
+                criteria.andLike("caption", "%" + searchMap.get("caption") + "%");
             }
             // 图片
-            if(searchMap.get("image")!=null && !"".equals(searchMap.get("image"))){
-                criteria.andLike("image","%"+searchMap.get("image")+"%");
+            if (searchMap.get("image") != null && !"".equals(searchMap.get("image"))) {
+                criteria.andLike("image", "%" + searchMap.get("image") + "%");
             }
             // 图片列表
-            if(searchMap.get("images")!=null && !"".equals(searchMap.get("images"))){
-                criteria.andLike("images","%"+searchMap.get("images")+"%");
+            if (searchMap.get("images") != null && !"".equals(searchMap.get("images"))) {
+                criteria.andLike("images", "%" + searchMap.get("images") + "%");
             }
             // 售后服务
-            if(searchMap.get("saleService")!=null && !"".equals(searchMap.get("saleService"))){
-                criteria.andLike("saleService","%"+searchMap.get("saleService")+"%");
+            if (searchMap.get("saleService") != null && !"".equals(searchMap.get("saleService"))) {
+                criteria.andLike("saleService", "%" + searchMap.get("saleService") + "%");
             }
             // 介绍
-            if(searchMap.get("introduction")!=null && !"".equals(searchMap.get("introduction"))){
-                criteria.andLike("introduction","%"+searchMap.get("introduction")+"%");
+            if (searchMap.get("introduction") != null && !"".equals(searchMap.get("introduction"))) {
+                criteria.andLike("introduction", "%" + searchMap.get("introduction") + "%");
             }
             // 规格列表
-            if(searchMap.get("specItems")!=null && !"".equals(searchMap.get("specItems"))){
-                criteria.andLike("specItems","%"+searchMap.get("specItems")+"%");
+            if (searchMap.get("specItems") != null && !"".equals(searchMap.get("specItems"))) {
+                criteria.andLike("specItems", "%" + searchMap.get("specItems") + "%");
             }
             // 参数列表
-            if(searchMap.get("paraItems")!=null && !"".equals(searchMap.get("paraItems"))){
-                criteria.andLike("paraItems","%"+searchMap.get("paraItems")+"%");
+            if (searchMap.get("paraItems") != null && !"".equals(searchMap.get("paraItems"))) {
+                criteria.andLike("paraItems", "%" + searchMap.get("paraItems") + "%");
             }
             // 是否上架
-            if(searchMap.get("isMarketable")!=null && !"".equals(searchMap.get("isMarketable"))){
-                criteria.andLike("isMarketable","%"+searchMap.get("isMarketable")+"%");
+            if (searchMap.get("isMarketable") != null && !"".equals(searchMap.get("isMarketable"))) {
+                criteria.andLike("isMarketable", "%" + searchMap.get("isMarketable") + "%");
             }
             // 是否启用规格
-            if(searchMap.get("isEnableSpec")!=null && !"".equals(searchMap.get("isEnableSpec"))){
-                criteria.andLike("isEnableSpec","%"+searchMap.get("isEnableSpec")+"%");
+            if (searchMap.get("isEnableSpec") != null && !"".equals(searchMap.get("isEnableSpec"))) {
+                criteria.andLike("isEnableSpec", "%" + searchMap.get("isEnableSpec") + "%");
             }
             // 是否删除
-            if(searchMap.get("isDelete")!=null && !"".equals(searchMap.get("isDelete"))){
-                criteria.andLike("isDelete","%"+searchMap.get("isDelete")+"%");
+            if (searchMap.get("isDelete") != null && !"".equals(searchMap.get("isDelete"))) {
+                criteria.andLike("isDelete", "%" + searchMap.get("isDelete") + "%");
             }
             // 审核状态
-            if(searchMap.get("status")!=null && !"".equals(searchMap.get("status"))){
-                criteria.andLike("status","%"+searchMap.get("status")+"%");
+            if (searchMap.get("status") != null && !"".equals(searchMap.get("status"))) {
+                criteria.andLike("status", "%" + searchMap.get("status") + "%");
             }
 
             // 品牌ID
-            if(searchMap.get("brandId")!=null ){
-                criteria.andEqualTo("brandId",searchMap.get("brandId"));
+            if (searchMap.get("brandId") != null) {
+                criteria.andEqualTo("brandId", searchMap.get("brandId"));
             }
             // 一级分类
-            if(searchMap.get("category1Id")!=null ){
-                criteria.andEqualTo("category1Id",searchMap.get("category1Id"));
+            if (searchMap.get("category1Id") != null) {
+                criteria.andEqualTo("category1Id", searchMap.get("category1Id"));
             }
             // 二级分类
-            if(searchMap.get("category2Id")!=null ){
-                criteria.andEqualTo("category2Id",searchMap.get("category2Id"));
+            if (searchMap.get("category2Id") != null) {
+                criteria.andEqualTo("category2Id", searchMap.get("category2Id"));
             }
             // 三级分类
-            if(searchMap.get("category3Id")!=null ){
-                criteria.andEqualTo("category3Id",searchMap.get("category3Id"));
+            if (searchMap.get("category3Id") != null) {
+                criteria.andEqualTo("category3Id", searchMap.get("category3Id"));
             }
             // 模板ID
-            if(searchMap.get("templateId")!=null ){
-                criteria.andEqualTo("templateId",searchMap.get("templateId"));
+            if (searchMap.get("templateId") != null) {
+                criteria.andEqualTo("templateId", searchMap.get("templateId"));
             }
             // 运费模板id
-            if(searchMap.get("freightId")!=null ){
-                criteria.andEqualTo("freightId",searchMap.get("freightId"));
+            if (searchMap.get("freightId") != null) {
+                criteria.andEqualTo("freightId", searchMap.get("freightId"));
             }
             // 销量
-            if(searchMap.get("saleNum")!=null ){
-                criteria.andEqualTo("saleNum",searchMap.get("saleNum"));
+            if (searchMap.get("saleNum") != null) {
+                criteria.andEqualTo("saleNum", searchMap.get("saleNum"));
             }
             // 评论数
-            if(searchMap.get("commentNum")!=null ){
-                criteria.andEqualTo("commentNum",searchMap.get("commentNum"));
+            if (searchMap.get("commentNum") != null) {
+                criteria.andEqualTo("commentNum", searchMap.get("commentNum"));
             }
 
         }
